serverless-api/app/requirements.txt
fastapi
uvicorn
mangum
httpx
python-multipart
ably
python-dotenv
boto3
openai
requests